+++ verifiers/v1/runtimes/prime.py
@@ -10,13 +10,8 @@ from typing import Literal
 from pydantic_config import BaseConfig
 
 from verifiers.v1.errors import ProgramError
-from verifiers.v1.runtimes.base import (
-    _TUNNEL_LIMITER,
-    ProgramResult,
-    Runtime,
-    creation_limiter,
-    parse_gpu,
-)
+from verifiers.v1.runtimes.base import ProgramResult, Runtime, parse_gpu
+from verifiers.v1.runtimes.limiters import _TUNNEL_LIMITER, creation_limiter
 
 logger = logging.getLogger(__name__)
 
@@ -56,7 +51,7 @@ class PrimeConfig(BaseConfig):
     creates_per_min: int | None = None
     """Pace sandbox creation to this many per minute, enforced host-wide across every
     env-server worker process (None/<= 0 disables it). (Tunnel creation is limited separately
-    and globally — see base._TUNNEL_LIMITER.)"""
+    and globally — see limiters._TUNNEL_LIMITER.)"""
 
 
 class PrimeRuntime(Runtime):
